refactor(microservices): move 8 transport implementations into transports/ subfolder
Move grpc, kafka, mqtt, nats, rabbitmq, redis-streams, redis, and tcp
transport files (+ their test files) into src/transports/ per the
package-folder-structure standard. Update all internal import paths.
Public API re-export signatures in index.ts remain unchanged.
