+++ packages/microservices/src/index.ts
@@ -5,26 +5,26 @@ export {
   MessagePattern,
   ServerStreamPattern,
 } from './decorators.js';
-export { GrpcMicroserviceTransport, type GrpcMicroserviceTransportOptions } from './grpc-transport.js';
-export { KafkaMicroserviceTransport, type KafkaMicroserviceTransportOptions } from './kafka-transport.js';
+export { GrpcMicroserviceTransport, type GrpcMicroserviceTransportOptions } from './transports/grpc-transport.js';
+export { KafkaMicroserviceTransport, type KafkaMicroserviceTransportOptions } from './transports/kafka-transport.js';
 export { MicroservicesModule, createMicroservicesProviders } from './module.js';
 export { MicroserviceLifecycleService } from './service.js';
-export { MqttMicroserviceTransport, type MqttMicroserviceTransportOptions } from './mqtt-transport.js';
-export { NatsMicroserviceTransport, type NatsMicroserviceTransportOptions } from './nats-transport.js';
+export { MqttMicroserviceTransport, type MqttMicroserviceTransportOptions } from './transports/mqtt-transport.js';
+export { NatsMicroserviceTransport, type NatsMicroserviceTransportOptions } from './transports/nats-transport.js';
 export {
   RedisPubSubMicroserviceTransport,
   type RedisPubSubMicroserviceTransportOptions,
-} from './redis-transport.js';
+} from './transports/redis-transport.js';
 export {
   RedisStreamsMicroserviceTransport,
   type RedisStreamClientLike,
   type RedisStreamsMicroserviceTransportOptions,
-} from './redis-streams-transport.js';
+} from './transports/redis-streams-transport.js';
 export {
   RabbitMqMicroserviceTransport,
   type RabbitMqMicroserviceTransportOptions,
-} from './rabbitmq-transport.js';
+} from './transports/rabbitmq-transport.js';
 export * from './status.js';
-export { TcpMicroserviceTransport } from './tcp-transport.js';
+export { TcpMicroserviceTransport } from './transports/tcp-transport.js';
 export { MICROSERVICE } from './tokens.js';
 export type { Microservice, MicroserviceModuleOptions, MicroserviceTransport, Pattern, ServerStreamWriter } from './types.js';
