+++ packages/microservices/src/transports/redis-streams-transport.ts
@@ -1,4 +1,4 @@
-import type { MicroserviceTransport, TransportHandler } from './types.js';
+import type { MicroserviceTransport, TransportHandler } from '../types.js';
 
 interface StreamReadGroupResult {
   readonly id: string;
